requirements.txt
fastapi[standard]==0.115.4
supabase==2.10.0
python-dotenv==1.0.1